==> 4_NRGV_Data_Analysis_Reconciliation.ipynb <==
# NRGV Data Analyst Interview Project

## Data Reconciliation

### Future Work - Ideas

1) Could test for count of underscores in file name to chack for additional encodings<br>
2) Using LAT/LONG and Date, determine sunset/sunrise; progamaticaly remove records when the sun is not out (from sunset to sunrise).<br>
2a) Eliminate these records during the injest stage -Document time-delta vs lines of code vs reduced records.<br>
3) Automate the detection of new state in dictionary.

### Notebook Objectives:

Data Reconciliation:
<ul>
    <li>Verify values</li>

-----------------------------

### Libraries

import requests as req     # DOWNLOADING URL VIA URL
import zipfile             # DEALING WITH ZIP COMPRESSED FILE
from io import BytesIO     # READING FILE FROM BUFFER
import os                  # DIRECTORY CREATION
import pandas as pd        # PANDAS DATAFRAME
import datetime            # DATE AND TIME
from datetime import datetime
import numpy as np

### Variables

# SWSD_SUMMARY_DATA DAYTIME 
sumry_daytime_path = 'C://Users//bruce//Documents//Data//SWSD_SUMMARY_DATA//ALL_STATES_SWSD_SUMMARY_DATA_DAYTIME.csv'

# SWSD_SUMMARY_DATA DAYTIME_AGGREGATIONS
sumry_daytime_agg_path = 'C://Users//bruce//Documents//Data//SWSD_SUMMARY_DATA//ALL_STATES_SWSD_SUMMARY_DATA_DAYTIME_AGGREGATIONS.csv'

states_dict = {
    "Alabama":"al",
    "Arizona":"az",
    "California":"ca",
    "Colorado":"co",
    "Florida":"fl",
    "Maine":"me",
    "New Mexico":"nm",
    "Texas":"tx"    
}

# DEFINE COLUMN NAMES
col_names = 'State','DataType','Latitude','Longitude','PVType','CapacityMW','Power(MW)_SUM','Power(MW)_MAX',\
'Power(MW)_MEAN','TimeIntervalMin'

# READ DATA
sumry_daytime = pd.read_csv(sumry_daytime_path, parse_dates=['LocalTime'])
sumry_daytime_agg = pd.read_csv(sumry_daytime_agg_path, header=0,names=col_names,skiprows=1)

-------------------------------------------------

### ALL_STATES_SWSD_SUMMARY_DATA_DAYTIME

sumry_daytime.head()

sumry_daytime.dtypes

# FILTERED FOR: STATE==AL

x = sumry_daytime[(sumry_daytime['State']=='AL')]['Power(MW)'].sum()

print("The Actual power produced in AL was: {:,.2f}".format(x))

# FILTERED FOR: STATE==AL, DataType=='Actual'

x = sumry_daytime[(sumry_daytime['State']=='AL') & (sumry_daytime['DataType']=='Actual')]['Power(MW)'].sum()

print("The Actual power produced in AL was: {:,.2f}".format(x))

# FILTERED FOR: STATE==AL, DataType=='Actual',  PVType=='UPV'

x = sumry_daytime[(sumry_daytime['State']=='AL') & (sumry_daytime['DataType']=='Actual')\
                  & (sumry_daytime['PVType']=='UPV')]['Power(MW)'].sum()

print("The Actual power produced in AL was: {:,.2f}".format(x))

# FILTERED FOR: STATE==AL, Latitude between 31.87 and 34.4, Longitude between -102.51 and -86.7
# block["Age"].between(12, 15, inclusive = True) 

x = sumry_daytime[(sumry_daytime['State']=='AL') & (sumry_daytime['Latitude'].between(31.87,34.4, inclusive = True))\
                 & (sumry_daytime['Longitude'].between(-102.51,-86.7, inclusive = True))]['Power(MW)'].sum()

print("The Actual power produced in AL was: {:,.2f}".format(x))

# FILTERED FOR: STATE==AL, Latitude between 31.87 and 34.4, Longitude between -102.51 and -86.7
# block["Age"].between(12, 15, inclusive = True) 

x = sumry_daytime[(sumry_daytime['State']=='AL') & (sumry_daytime['Latitude'] == 33.75)\
                 & (sumry_daytime['Longitude']== -86.75)]['Power(MW)'].sum()

print("The Actual power produced in AL was: {:,.2f}".format(x))

# FILTERED FOR: DataType=='Actual'

x = sumry_daytime[(sumry_daytime['DataType']=='Actual')]['Power(MW)'].max()

print("The Max Power(MW) produced was: {:,.2f}".format(x))
#print(sumry_daytime[(sumry_daytime['DataType']=='Actual')] & (sumry_daytime['Power(MW)']==x))
sumry_daytime[sumry_daytime['Power(MW)']==x]

------------------------------------------------------------------------------------

### ALL_STATES_SWSD_SUMMARY_DATA_DAYTIME_AGGREGATIONS 

sumry_daytime_agg.head()

# FILTERED FOR: STATE==AL, Latitude between 31.87 and 34.4, Longitude between -102.51 and -86.7
# block["Age"].between(12, 15, inclusive = True) 

x = sumry_daytime_agg[(sumry_daytime_agg['State']=='AL') & \
                      (sumry_daytime_agg['Latitude'].between(31.87,34.4, inclusive = True))\
                      & (sumry_daytime_agg['Longitude'].between(-102.51,-86.7, inclusive = True))\
                      & ()]['Power(MW)_SUM'].sum()

print("The Actual power produced in AL was: {:,.2f}".format(x))
